Set up grid gaps and 1st and last column
column on outside of content.

Jira ticket: CAMS-219
