+++ user-interface/src/App.scss
@@ -10,7 +10,7 @@
   padding: 0;
 
   .body {
-    padding: 0 50px 50px 50px;
+    padding: 0 0 50px 0;
 
     h1, h2 {
       text-align: left;
