+++ user-interface/src/case-assignment/CaseAssignmentScreen.scss
